Refactored MusicUtil
